% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppIsomap-package.R
\docType{package}
\name{cppIsomap-package}
\alias{cppIsomap-package}
\title{cppIsomap}
\description{
cppIsomap: RCpp port of RDRToolbox's Isomap embedding implementation
}
\details{
Provides cppIsomap function that computes the Isomap embedding as introduced in 2000 by Tenenbaum, de Silva and Langford. 
Direct port of RDRToolbox version. Does not implement residual plot.

The only function exported by this package is cppIsomap, which has the same parameters as RDRToolbox::Isomap.
}
\author{
Marc Burgess
}
\keyword{internal}
