% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppIsomap}
\alias{cppIsomap}
\title{Computes the Isomap embedding as introduced in 2000 by Tenenbaum, de Silva and Langford.}
\usage{
cppIsomap(data, dims = as.integer(c(2)), k = NULL, mod = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{N x D matrix (N samples, D features)}

\item{dims}{vector containing the target space dimension(s)}

\item{k}{number of neighbours}

\item{mod}{use modified Isomap algorithm}

\item{verbose}{show a summary of the embedding procedure at the end}
}
\value{
It returns a N x dim matrix (N samples, dim features) with the reduced input data (list of several matrices if more than one dimension was specified)
}
\description{
Isomap is a nonlinear dimension reduction technique, that preserves global properties of the data. 
That means, that geodesic distances between all samples are captured best in the low dimensional embedding. 
This C++ version is based R version by Christoph Bartenhagen, which is in turn based on the Matlab implementation by Tenenbaum and uses Floyd's Algorithm to compute the neighbourhood graph of shortest distances, when calculating the geodesic distances.
A modified version of the original Isomap algorithm is included. It respects nearest and farthest neighbours.
}
\examples{
\dontrun{
## two dimensional Isomap embedding of a 1.000 dimensional dataset using k=5 neighbours
## two dimensional Isomap embedding of a 1.000 dimensional dataset using k=5 neighbours
d = RDRToolbox::generateData(samples=20, genes=1000, diffgenes=100, blocksize=10)
d_low = Isomap(data=d[[1]], dims=2, k=5)
## Isomap residuals for target dimensions 1-10
d_low = Isomap(data=d[[1]], dims=1:10, k=5, plotResiduals=TRUE)	
    
## three dimensional Isomap embedding of a 1.000 dimensional dataset using k=10 (nearest and farthest) neighbours
d = generateData(samples=20, genes=1000, diffgenes=100, blocksize=10)
d_low = Isomap(data=d[[1]], dims=3, mod=TRUE, k=10)
}
}
\seealso{
\link[RDRToolbox]{Isomap}
}
